EN-2248 Rename encryption_utils to crypto_utils.
