+++ incountry/__init__.py
@@ -1,5 +1,5 @@
 from .storage import Storage
 from .incountry_crypto import InCrypto, InCryptoException
-from .encryption_utils import decrypt_record, encrypt_record, get_salted_hash
+from .crypto_utils import decrypt_record, encrypt_record, get_salted_hash
 from .secret_key_accessor import SecretKeyAccessor
 from .exceptions import SecretKeyAccessorException, StorageError, StorageClientError, StorageServerError
